Added the Send button.
Added the 'Send' button as well as messagebox to get EmoJiID selected
and the Total number of coins selected. Player can select and deselect
the coins before pressing the send button, however, in case we decide to
add the 'reset' button, we already have the code to move the coins back
to the original places.